requirements/prod.txt
Django==1.6.5
djangorestframework==2.3.14
drf-compound-fields==0.2.1

-e git+https://github.com/clouserw/tower.git@ce2eba049a5146918fa5239d827bdde9e36eadb7#egg=tower-dev
